requirements.txt
alembic==1.15.1
fastapi==0.115.12
fastapi_mail==1.4.2
passlib==1.7.4
pydantic==2.11.3
pydantic_settings==2.8.1
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
bcrypt==4.3.0
mercadopago==2.3.0
pymysql==1.1.0
uvicorn==0.23.2
python-multipart==0.0.6